requirements.txt
matplotlib==3.0.3
numpy==1.16.2
pyyaml==5.1.2
sacrebleu==1.4.3
scipy==1.2.1
torch==1.4.0
tqdm==4.30.0
